# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Opora API
  version: 0.3.0
  description: Authenticated API for the Opora student document workspace.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/auth/invitations/accept:
    post:
      operationId: acceptInvitation
      tags: [Auth]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/AcceptInvitation" } } }
      responses:
        "200": { description: Invitation accepted and password set. }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "410": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: opora_session }
    csrfHeader: { type: apiKey, in: header, name: X-CSRF-Token }
...
  responses:
...
    Error:
      description: Stable public error envelope.
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
  schemas:
...
    AcceptInvitation:
      type: object
      additionalProperties: false
      required: [token, password]
      properties:
        token: { type: string, minLength: 32 }
        password: { type: string, format: password, minLength: 12 }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }